Build basic rollup workflow to bundle js
For long term benefit, it is good to use a bundling tool to produce multiple versions of js code. I target to provide:

1. ES6 module aka `import Validid from 'validid'` That's why I recommend to use node 13 although it is not in LTS yet.
2. CommonJs aka `var Validid = require('validid')`
3. IIFE code that supports legacy browsers. It runs `window.validid`.
4. UMD to support multiple uses.

In <package.json>:

- "main" serves flattened UMD codes, so developers can import it in different ways.
- "browser" serves the flattened IIFE codes. The code is transpiled (inc. `class`) so all browsers can run it directory.
- File name in "main" of <package.json>: <validid.umd.js> (temporary decision)
- File name in "browser": <validid.browser.js> (temporary decision)

"browser" is not a good name because developers will take it for their own front-end workflows. However, existing file is for end-use in browsers. It is subjected to change.